gate: replace capability-preflight with per-test self-detection
Owner ruling: the classify-once-then-skip-regex preflight is deleted; the
idiomatic Go mechanism takes its place. New internal/e2ecap package
provides RequireLoopbackBind/RequireProcessInspect — the probe runs at
test time and t.Skips with the reason when the host lacks the capability.
Applied to the exact families the preflight's skip pattern covered: all
17 cmd/evener-hub TestE2E_* tests and the 22 cmd/evener-tui tmux call
sites (agent's unrelated TestE2E_* tests were never covered and are
untouched).

Deleted: cmd/evener-dev capability-preflight, internal/devtool/
capabilityprobe, gatesurface's CapabilitySkipPattern and its drift pin,
gate-surface-lib.sh's skip registry, run-module-tests.sh's
EVENER_GATE_CAPABILITY_SKIP plumbing, and merge-approval-gate's
issue-#181 export dance (the target is now a plain four-phase chain).
The selftests' capability checks go with them; phase-order and
make-wiring checks stay.

Verified: make lint, make test-dev-tooling, root+agent short suites, and
ROOT_FULL=1 WEB=0 make test all green; a hub e2e test runs (not skips) on
an unrestricted host.
