+++ cmd/evener-dev/main.go
@@ -15,9 +15,8 @@ import (
 )
 
 var subcommands = map[string]func(args []string) int{
-	"agent-shards":         runAgentShards,
-	"module-lint":          lintMain,
-	"capability-preflight": runCapabilityPreflight,
+	"agent-shards": runAgentShards,
+	"module-lint":  lintMain,
 }
 
 func main() {
