+++ cmd/evener-hub/e2e_startup_context_test.go
@@ -9,6 +9,8 @@ import (
 	"testing"
 	"time"
 
+	"primeradiant.com/evener/internal/e2ecap"
+
 	"primeradiant.com/evener/appwire"
 	"primeradiant.com/evener/test/e2e/fakellm"
 )
@@ -31,6 +33,8 @@ import (
 // daemon's turn in it. The context still has to reach the model, riding the
 // user's first turn the way steering is meant to.
 func TestE2E_StartingASessionThatHasDaemonStartupContext(t *testing.T) {
+	e2ecap.RequireLoopbackBind(t)
+	e2ecap.RequireProcessInspect(t)
 	if testing.Short() {
 		t.Skip("live-stack e2e: builds binaries and runs a hub + daemon")
 	}
