+++ cmd/evener-hub/e2e_turn_control_test.go
@@ -17,6 +17,8 @@ import (
 	"testing"
 	"time"
 
+	"primeradiant.com/evener/internal/e2ecap"
+
 	"primeradiant.com/evener/appwire"
 	"primeradiant.com/evener/identifier"
 	"primeradiant.com/evener/test/e2e/fakellm"
@@ -40,6 +42,8 @@ import (
 // no credential, no network (AGENTS.md: evener plumbing gets a scripted
 // provider; only model behaviour stays live).
 func TestE2E_TurnControlReachesTheSession(t *testing.T) {
+	e2ecap.RequireLoopbackBind(t)
+	e2ecap.RequireProcessInspect(t)
 	if testing.Short() {
 		t.Skip("live-stack e2e: builds binaries and runs a hub + daemon")
 	}
@@ -183,6 +187,8 @@ func TestE2E_TurnControlReachesTheSession(t *testing.T) {
 // turn_m<n>, so steer and stop came back Conflict("turn is not active") and
 // the composer showed nothing at all.
 func TestE2E_TurnControlReachesAnAgentStartedTurn(t *testing.T) {
+	e2ecap.RequireLoopbackBind(t)
+	e2ecap.RequireProcessInspect(t)
 	if testing.Short() {
 		t.Skip("live-stack e2e: builds binaries and runs a hub + daemon")
 	}
@@ -302,6 +308,8 @@ func TestE2E_TurnControlReachesAnAgentStartedTurn(t *testing.T) {
 // boundary lands (kata c2ty keeps that window open deliberately), and steering
 // against that id is the bug, not the test.
 func TestE2E_TurnControlReachesANotificationTurn(t *testing.T) {
+	e2ecap.RequireLoopbackBind(t)
+	e2ecap.RequireProcessInspect(t)
 	if testing.Short() {
 		t.Skip("live-stack e2e: builds binaries and runs a hub + daemon")
 	}
