+++ scripts/lib/gate-surface-lib.sh
@@ -23,42 +23,3 @@ GATE_FUZZ_TEST_SKIP='(SeqFuzz|SchemaFuzz|Structured.*Reach|LifecycleAdapter|Tool
 # which is make fuzz's job, not the default gate's.
 GATE_TEST_RUN='^(Test|Example)'
 
-# --- sandbox capability skip registry (kata 5gvk) ---------------------------
-#
-# evener-dev capability-preflight classifies four sandbox-sensitive host
-# capabilities ONCE, up front: loopback binds, a Chrome/Chromium binary,
-# process inspection via `ps`, and a writable external git cache directory.
-# Any it finds blocked feeds this registry to decide which KNOWN test-name
-# pattern to skip instead of letting those tests fail into a denied
-# bind/exec/write. gate_capability_skip_pattern CAPABILITY_ID prints the
-# pattern for one id, or nothing when that capability has no known consumer.
-#
-# Evidenced mapping, root module ONLY (kata 5gvk's premise check):
-#   - loopback-bind, process-inspect: cmd/evener-hub's TestE2E_* family and
-#     cmd/evener-tui's TestTUITmuxE2E_* family are the root module's ONLY test
-#     files that both (a) run only under ROOT_FULL=1 (gated by
-#     testing.Short(); merge-approval-gate is the only caller that sets it)
-#     and (b) spawn a real hub/daemon or tmux pane bound to a real loopback
-#     port. Ordinary `make test` never reaches them.
-#   - chrome-cdp, git-cache: no test file anywhere in this tree consumes
-#     either today. test-web-browser needs Chrome but is a separate,
-#     non-gate target; the only thing in the tree that names the fixed
-#     /tmp/git-cache path is internal/devtool/capabilityprobe's own default,
-#     which the probe creates in order to prove it is writable. Both are probed
-#     and reported for completeness and honesty; the pattern is empty because
-#     nothing is skipped yet.
-#
-# Applied to the ROOT module ONLY, deliberately: `agent/session_escalation_e2e_test.go`
-# (Linux-only, unrelated sandbox-escalation coverage) also has two
-# `TestE2E_*`-named tests that share nothing with the cmd/evener-hub family
-# except the prefix. Unioning this pattern into every module's `-skip` would
-# silently and wrongly skip those two whenever loopback-bind or
-# process-inspect is blocked. run-module-tests.sh applies the pattern to "."
-# only for exactly this reason - see its root_skip.
-gate_capability_skip_pattern() {
-	case "$1" in
-		loopback-bind | process-inspect) printf '%s' '^(TestE2E_|TestTUITmuxE2E_)' ;;
-		chrome-cdp | git-cache) printf '' ;;
-		*) printf '' ;;
-	esac
-}
